Revert "WIP: Push Notifications"
This reverts commit ea36a382fc78f0d8ad1a84ff3e23466d52f55ff4.
