+++ example/android/.gitignore
@@ -5,5 +5,3 @@ gradle-wrapper.jar
 /gradlew.bat
 /local.properties
 GeneratedPluginRegistrant.java
-
-app/google-services.json
