--- example/lib/op_state.dart
@@ -1,6 +1,0 @@
-enum OpState {
-  notStarted,
-  inProgress,
-  succeeded,
-  failed,
-}
