+++ lib/src/platform/src/rest/rest.dart
@@ -30,7 +30,6 @@ class Rest extends PlatformObject implements RestInterface<RestChannels> {
         options = options ?? ClientOptions.fromKey(key!),
         super() {
     channels = RestChannels(this);
-    push = PushNative(handle);
   }
 
   @override
@@ -83,7 +82,7 @@ class Rest extends PlatformObject implements RestInterface<RestChannels> {
   }
 
   @override
-  late Push push;
+  Push? push;
 
   @override
   late RestChannels channels;
